Remove obsolete GUI code and stale docs
