--- gui/BrakeDiscInspector_GUI_ROI/desktop.ini
@@ -1,2 +1,0 @@
-[LocalizedFileNames]
-BackendAPI.cs=@BackendAPI.cs,0
